fix bug in kwargs_to_matrix and test
