+++ tests/test_scenegraph.py
@@ -1,0 +1,89 @@
+try:
+    from . import generic as g
+except BaseException:
+    import generic as g
+
+from trimesh.scene.transforms import EnforcedForest
+
+
+def random_chr():
+    return chr(ord('a') + int(round(g.np.random.random() * 25)))
+
+
+class GraphTests(g.unittest.TestCase):
+
+    def test_forest(self):
+        g = EnforcedForest()
+        for i in range(5000):
+            g.add_edge(random_chr(), random_chr())
+
+    def test_cache(self):
+        for i in range(10):
+            scene = g.trimesh.Scene()
+            scene.add_geometry(g.trimesh.creation.box())
+
+            mod = [scene.graph.modified()]
+            scene.set_camera()
+            mod.append(scene.graph.modified())
+            assert mod[-1] != mod[-2]
+
+            assert not g.np.allclose(
+                scene.camera_transform,
+                g.np.eye(4))
+            scene.camera_transform = g.np.eye(4)
+            mod.append(scene.graph.modified())
+            assert mod[-1] != mod[-2]
+
+            assert g.np.allclose(
+                scene.camera_transform,
+                g.np.eye(4))
+            assert mod[-1] != mod[-2]
+
+    def test_successors(self):
+        s = g.get_mesh('CesiumMilkTruck.glb')
+        assert len(s.graph.nodes_geometry) == 5
+
+        # world should be root frame
+        assert (s.graph.transforms.successors(
+            s.graph.base_frame) == s.graph.nodes)
+
+        for n in s.graph.nodes:
+            # successors should always return subset of nodes
+            succ = s.graph.transforms.successors(n)
+            assert succ.issubset(
+                s.graph.nodes)
+            # we self-include node in successors
+            assert n in succ
+
+        # test getting a subscene from successors
+        ss = s.subscene('3')
+        assert len(ss.geometry) == 1
+        assert len(ss.graph.nodes_geometry) == 1
+
+    def test_kwargs(self):
+        # test the function that converts various
+        # arguments into a homogenous transformation
+        f = g.trimesh.scene.transforms.kwargs_to_matrix
+        # no arguments should be an identity matrix
+        assert g.np.allclose(f(), g.np.eye(4))
+
+        # a passed matrix should return immediately
+        fix = g.np.random.random((4, 4))
+        assert g.np.allclose(f(matrix=fix), fix)
+
+        quat = g.trimesh.unitize([1, 2, 3, 1])
+        trans = [1.0, 2.0, 3.0]
+        rot = g.trimesh.transformations.quaternion_matrix(quat)
+        # should be the same as passed to transformations
+        assert g.np.allclose(rot, f(quaternion=quat))
+
+        # try passing both quaternion and translation
+        combine = f(quaternion=quat, translation=trans)
+        # should be the same as passed and computed
+        assert g.np.allclose(combine[:3, :3], rot[:3, :3])
+        assert g.np.allclose(combine[:3, 3], trans)
+
+
+if __name__ == '__main__':
+    g.trimesh.util.attach_to_log()
+    g.unittest.main()
